pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "uncertainty-lab"
version = "0.1.0"
description = "Local Uncertainty Lab: binary image classification with uncertainty metrics"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  "transformers>=4.36.0",
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "PyYAML>=6.0",
  "matplotlib>=3.7.0",
  "Pillow>=9.0.0",
]

[project.optional-dependencies]
ui = ["streamlit>=1.28.0"]

[project.scripts]
uncertainty-lab = "uncertainty_lab.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["uncertainty_lab*", "models*"]
